add cancellations

syntax errors


use std::vector instead of vector in cut_solver temporarily


fix in is_upper_bound, is_lower_bound


add bound() for polynomial, needs more testing


work on resolve


implement resolve()


implement improves()


replace low_bound by lower_bound


work on cut_solver


work on cut_solver


better printing in cut_solver.h


add value vector to cut_solver


work on bound propagaion for cut_solver


bound propagation for integer inequalites


bound propagation for integer inequalites


bound propagattions on integers


adding m_explanation field to cut_solver


simplify bound propagation in cut_solver


calculate conflict explanation


change m_explanation type to a set


making cut_solver a member of int_solver, missing push/pop support


Nikolaj's comments


Nikolaj's comments


return explanations from cut_solver and hook up push/pop


hook up push/pop


using resize of std::vector


it is a big squashed commit


rename hpp to cpp


fixes in push/pop of cut_solver


return simple inequalities a part of a conflict


work on propagation and the main loop


add file


work on cut_solver


init m_v[j], the var values only when j is fixed


cut_solver


handle decide in cut_solver


start on resolve_conflict


start on resolve_conflict


remove cut_solver_def.h


in the middle


change signature of resolve


fix the domain of the decided var


work on tightening of ineqs


work on tight ineqs


work on tightening


work on tightening


resolve conflict


fix in usage of resolve()


work on conflict resolution


cut_solver is not a template


represent var_info as a class, not a struct


make literal a class


better resolve_conflict scheme, and switch to *constraints in literals


debug conflict resolution in cut_solver


switch to vector


remove nondetermenistic behavior from cut_solver


debug resolve conflict


fix backjump


fix backjump


fix backjump


fix backjump


dumb explanation


get rid of a parameter


add lemmas origins


use lemma_origins to provide correct explanations


use lemma_origins to provide correct explanations


store lemmas in a separate vector


use std::unordered_set for m_dependent_constraints


use std::unordered_set for m_dependent_constraints


fix bugs with lemmas


finding conflicting cores


switch from changed variables to active_set


less active constraints


work on cut_solver.h


propagate simple constraing immediately


propagate simple constraints immediately


fixing bugs with active set


remove const_cast


towards unbounded variables


toward unbounded variables


toward unbounded variables


make lemmas_origins a set


use correct hash and equal in m_lemma_origins


add testing code


add testing code


debug


debug unlimited vars


take in Nikolaj's comments and improvements


address the comments


handle unlimited vars in check_inconsistent


debug


detect trivial polynomials in resolve


Nikolaj's changes


simplify handling of m_global_bound_var


decide on m_global_bound_var if it is not fixed


simplify m_global_bound_var


remove m_global_bound_var, simplify the indexing of var_infos of cut_solver


do not run cut_solver with vars without any bound


small changes


add cancellation in cut_solver


do not pop lemmas during a cut_solver run


treating cut_solver as an heurisitic


prepare for cut_solver returning undef


simplify work with active_set in cut_solver, add stats


simplify var_info literals


fix a bug in fill_conflict_explanation


fix a bug in the conflict explanation


add timeout to validate_* in theory_lra.cpp


simplify cut_solver, no special treatment for simple constraints


cleanup the cancel story


cleanup cancelling


fix a bug in push/pop of cut_solver


extract a method in int_solver


some progress with the new scheme


add testing code


fixes in test and in literal creation


fix a bug in bound propagation in cut_solver.h


simplify cut_solver


provide valid conflict explanation


use a lazy push in stacked_map


use a lazy push in stacked_map


optimize stack operations on var_info's domains


fix a bug in tightening


use the standard tactics from qflia_tactic


spread the var domain stack over literals


spread the var domain stack over literals


avoid cycling in cut_solver.h and fixes in push/pop


fixes after rebase

+++ src/smt/theory_str.cpp
@@ -11074,8 +11074,8 @@ namespace smt {
                 ctx.force_phase(lit);
             }
 
-			zstring aStr = gen_val_string(len, options[static_cast<int>(i) - static_cast<int>(l)]);
-			expr * strAst;
+            zstring aStr = gen_val_string(len, options[i - l]);
+            expr * strAst;
             if (m_params.m_UseFastValueTesterCache) {
                 if (!valueTesterCache.find(aStr, strAst)) {
                     strAst = mk_string(aStr);
